add --help -h to various scripts

+++ root/bin/auth0/stop-auth0.sh
@@ -2,6 +2,10 @@
 # make sure windows git bash does not alter paths
 export MSYS_NO_PATHCONV=1
 
-docker rm -f "$AUTH0_CONTAINER_NAME"
+if [[ "$1" = "--help" || "$1" = "-h" ]]; then
+  echo "Used to stop running auth0 container with name $AUTH0_CONTAINER_NAME"
+  exit 0
+fi
 
-sleep 2
+
+docker rm -f "$AUTH0_CONTAINER_NAME"
